more house keeping
working, needs more cleaning.

need to test modbus writing next
